refactor: some clean up
